requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.4.4
pymongo==4.4.0
python-dotenv==1.0.0
werkzeug==3.0.0
Flask-Cors==4.0.0
